Release 11

+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface FilmStorage {
     Film create(Film film); // Создание нового фильма
 
-    Optional<Film> getFilmById(int id); // Получение фильма по ID
+    Optional<Film> getFilmById(Long id); // Получение фильма по ID
 
     Collection<Film> findAll(); // Получение всех фильмов
